src/roommatefinder/requirements/_base.txt
Django~=3.0.3
psycopg2-binary==2.9.9
pytz==2019.2
sqlparse==0.3.0
djangorestframework==3.15.1
django-multiselectfield==0.1.12
django-model-utils==4.5.0
